Rewrite DatabaseHandler for read-only database
This removes the dependence on the sqliteassethelper. We have a
read-only database anyway, so we never need DB updates.
Instead, if the database does not exist, a database error occurs, or
the database version does not correspond to the app version, we just
copy the whole database from assets to the data dir, overwriting any old
database.